rename not equals, equals nodes
